# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Blog",
        "contact": {
            "email": "admin@example.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "https://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost/api",
            "description": "Laravel Swagger API Server"
        }
    ],
    "paths": {
...
        "/user": {
            "get": {
                "tags": [
                    "User"
                ],
                "summary": "Show user",
                "description": "Show user",
                "operationId": "9eeadb345322a9a5e6376e4019e3f3d5",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "firstname": {
                                        "type": "string",
                                        "example": "Barry"
                                    },
                                    "lastname": {
                                        "type": "string",
                                        "example": "Berkman"
                                    },
                                    "password": {
                                        "type": "string",
                                        "example": "asflkc42212"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "$ref": "#/components/schemas/User"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation error"
                    },
                    "404": {
                        "description": "not found"
                    },
                    "401": {
                        "description": "unauthorized"
                    }
                }
            },
            "put": {
                "tags": [
                    "User"
                ],
                "summary": "Update a user",
                "description": "Update a user",
                "operationId": "3f3f92809b94d65d5239eb23fb45e047",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "firstname": {
                                        "type": "string",
                                        "example": "Barry"
                                    },
                                    "lastname": {
                                        "type": "string",
                                        "example": "Berkman"
                                    },
                                    "password": {
                                        "type": "string",
                                        "example": "asflkc42212"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "$ref": "#/components/schemas/User"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation error"
                    },
                    "404": {
                        "description": "not found"
                    },
                    "401": {
                        "description": "unauthorized"
                    }
                }
            }
        },
        "/likes": {
            "get": {
                "tags": [
                    "User"
                ],
                "summary": "Show liked posts",
                "description": "List of user`s likes",
                "operationId": "cf2d4b8c65ee2ecca40fc56b9167e880",
                "responses": {
                    "201": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "description": "List of liked posts",
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/UserLike"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation error"
                    },
                    "404": {
                        "description": "not found"
                    },
                    "401": {
                        "description": "unauthorized"
                    }
                }
            }
        },
        "/comments": {
            "get": {
                "tags": [
                    "User"
                ],
                "summary": "Show commented posts",
                "description": "List of user comments",
                "operationId": "6ad0c06286abcdf936454cf9d4b19da9",
                "responses": {
                    "201": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "description": "List of commented posts",
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/UserComment"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation error"
                    },
                    "404": {
                        "description": "not found"
                    },
                    "401": {
                        "description": "unauthorized"
                    }
                }
            }
        }
    },
    "components": {
        "schemas": {
...
            "PostList": {
                "required": [
                    "id",
                    "title",
                    "text",
                    "category",
                    "comments_count",
                    "likes_count"
                ],
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "title": {
                        "type": "string",
                        "example": "group foto"
                    },
                    "text": {
                        "type": "string",
                        "example": "head too heavy"
                    },
                    "category": {
                        "type": "string",
                        "example": "summer"
                    },
                    "comments_count": {
                        "type": "integer",
                        "example": 1
                    },
                    "likes_count": {
                        "type": "integer",
                        "example": 1
                    }
                },
                "type": "object"
            },
...
            "User": {
                "required": [
                    "firstname",
                    "lastname",
                    "login"
                ],
                "properties": {
                    "firstname": {
                        "type": "string",
                        "example": "Barry"
                    },
                    "lastname": {
                        "type": "string",
                        "example": "Berkman"
                    },
                    "login": {
                        "type": "string",
                        "example": "bberkman@gmail.com"
                    }
                },
                "type": "object"
            },
            "UserComment": {
                "required": [
                    "id",
                    "text",
                    "user_firstname",
                    "user_lastname",
                    "link"
                ],
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "text": {
                        "type": "string",
                        "example": "That`s awesome"
                    },
                    "user_firstname": {
                        "type": "string",
                        "example": "Walter"
                    },
                    "user_lastname": {
                        "type": "string",
                        "example": "White"
                    },
                    "link": {
                        "type": "string",
                        "example": "http://localhost/api/posts/1"
                    }
                },
                "type": "object"
            },
            "UserLike": {
                "required": [
                    "user_id",
                    "post"
                ],
                "properties": {
                    "user_id": {
                        "type": "integer",
                        "example": 1
                    },
                    "post": {
                        "$ref": "#/components/schemas/PostList"
                    }
                },
                "type": "object"
            }
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "name": "bearerAuth",
                "in": "header",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    }
}